Moved folder and added skeleton for code challenge, including test skeleton.
