+++ javascript/code-challenges/linked-list/__tests__/linked-list.test.js
@@ -1,0 +1,20 @@
+'use strict';
+
+// Require our linked list class
+const LinkedList = require('../linked-list');
+
+// Write tests to prove the following functionality:
+
+//     Can successfully instantiate an empty linked list
+//     Can properly insert into the linked list
+//     The head property will properly point to the first node in the linked list
+//     Can properly insert multiple nodes into the linked list
+//     Will return true when finding a value within the linked list that exists
+//     Will return false when searching for a value in the linked list that does not exist
+//     Can properly return a collection of all the values that exist in the linked list
+
+describe('Linked List', () => {
+  it('exists', () => {
+    expect(true).toBeTruthy();
+  });
+});
