+++ javascript/code-challenges/linked-list/linked-list.js
@@ -1,0 +1,22 @@
+'use strict';
+
+class LinkedList {
+  insert(value) {
+    // Takes a value argument and adds a new node with that value to the head
+    return;
+  }
+  includes(value) {
+    // Takes a value argument and returns a boolean if that value exists
+    let existsResult = false;
+    // Traverse the linked list searching for the value.
+    return existsResult;
+  }
+  toString() {
+    // Returns a string representing all values in the linked list
+    let arrayString = '';
+    // Traverse the linked list and concat the values.
+    return arrayString;
+  }
+}
+
+module.exports = LinkedList;
