+++ javascript/code-challenges/linked-list/list-node.js
@@ -1,0 +1,10 @@
+'use strict';
+
+class ListNode {
+  constructor(){
+    this.value = 0;
+    this.next = '';
+  }
+}
+
+module.exports = ListNode;
