--- javascript/linked-list/index.js
@@ -1,7 +1,0 @@
-'use strict';
-
-class LinkedList {
-
-}
-
-module.exports = LinkedList;
